meta_generator/pyproject.toml
[tool.poetry]
name = "meta-generator"
version = "0.1.0"
description = ""
authors = ["RodionfromHSE <rodionkhvorostov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.73.0"
jsonschema = "^4.23.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
